third_party/blink/web_tests/external/wpt/media-source/mediasource-remove-expected.txt

This is a testharness.js-based test.
PASS Test remove with an negative start.
PASS Test remove with non-finite start.
PASS Test remove with a start beyond the duration.
PASS Test remove with a start larger than the end.
PASS Test remove with a NEGATIVE_INFINITY end.
PASS Test remove with a NaN end.
PASS Test remove after SourceBuffer removed from mediaSource.
PASS Test remove with a NaN duration.
PASS Test remove while update pending.
PASS Test aborting a remove operation.
PASS Test remove with a start at the duration.
PASS Test remove transitioning readyState from 'ended' to 'open'.
FAIL Test removing all appended data. assert_equals: Initial buffered range. expected "{ [0.112, 6.552) }" but got "{ [0.000, 6.552) }"
FAIL Test removing beginning of appended data. assert_equals: Initial buffered range. expected "{ [0.112, 6.552) }" but got "{ [0.000, 6.552) }"
FAIL Test removing the middle of appended data. assert_equals: Initial buffered range. expected "{ [0.112, 6.552) }" but got "{ [0.000, 6.552) }"
FAIL Test removing the end of appended data. assert_equals: Initial buffered range. expected "{ [0.112, 6.552) }" but got "{ [0.000, 6.552) }"
Harness: the test ran to completion.
